Ensure correct file length and sample skip calculations on 32-bit systems by using uint64_t
On 32-bit architectures, size_t may not be large enough to represent large file
sizes or sample counts. This commit updates the file length computation and the
return type of samplesToSkip() to use uint64_t, allowing accurate handling of
large files and ensuring consistency across platforms.

+++ src/algorithms/signal_source/adapters/file_source_base.h
@@ -112,10 +112,10 @@ protected:
     virtual double packetsPerSample() const;
 
     //! Compute the number of samples to skip
-    virtual size_t samplesToSkip() const;
+    virtual uint64_t samplesToSkip() const;
 
     //! Compute the number of samples in the file
-    size_t computeSamplesInFile() const;
+    uint64_t computeSamplesInFile() const;
 
     //! Abstracted front-end source. Sub-classes may override if they create specialized chains to
     //! decode source files into a usable format
